Rename and relocate
