+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli.Evaluations/Azure.Sdk.Tools.Cli.Evaluations.csproj
