+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli.Evaluations/ExpectedToolInputEvaluator.cs
@@ -9,7 +9,7 @@ using Microsoft.Extensions.AI.Evaluation;
 using Microsoft.Extensions.AI.Evaluation.Quality;
 using ModelContextProtocol.Protocol;
 
-namespace azsdk_mcp_evals
+namespace Azure.Sdk.Tools.Cli.Evaluations
 {
    public class ExpectedToolInputEvaluator : IEvaluator
    {
