+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli.Evaluations/ExpectedToolInputEvaluatorContext.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.AI;
 using Microsoft.Extensions.AI.Evaluation;
 
-namespace azsdk_mcp_evals
+namespace Azure.Sdk.Tools.Cli.Evaluations
 {
     public class ExpectedToolInputEvaluatorContext(IEnumerable<ChatMessage> messages)
         : EvaluationContext(name: ExpectedResultContextName, content: String.Join(",", messages))
